requirements.txt
asgiref==3.8.1
class-registry==2.1.2
Django==5.0.4
django-ckeditor-5==0.2.13
django-filter==24.2
djangorestframework==3.15.1
Faker==25.0.0
filters==1.3.2
pdf2image==1.17.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.28
six==1.16.0
sqlparse==0.5.0
